% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.ecdf.R
\name{calc.ecdf}
\alias{calc.ecdf}
\title{Calculated the empirical CDF of a time series}
\usage{
calc.ecdf(x, plot = TRUE)
}
\arguments{
\item{x}{a numeric vector representing the univarate time series}

\item{plot}{defiindicates if the ecdf should be plotted}
}
\value{
The function returns a ecdf function for the time series
}
\description{
Calculated the empirical CDF of a time series
}
