% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.eSAX.R
\name{create.eSAX}
\alias{create.eSAX}
\title{Perfom an eSAX transformation for time series}
\usage{
create.eSAX(x, breaks, w = 1000)
}
\arguments{
\item{x}{a numeric vector representing the univarate time series}

\item{breaks}{the breakpoints used for the SAX representation}

\item{w}{defines the word size used for SAX transformation}
}
\value{
The function returns an eSAX representation of x.
}
\description{
The function creates eSAX symbols for an univariate time series.
}
