% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.SAX.R
\name{create.SAX}
\alias{create.SAX}
\title{Perfom a SAX for time series}
\usage{
create.SAX(x, w, a, eps, norm, return.pieces = FALSE)
}
\arguments{
\item{x}{a numeric vector representing the univarate time series}

\item{w}{defines the word size used for SAX transformation}

\item{a}{defines the alphabet size used for SAX transformation}

\item{eps}{s the minimum threshold for variance in subsequence and should be
a numeric value. If the subsequence considered has a smaller variance than eps,
it will be represented as a word using the middle alphabet. The default value is 0.1}

\item{norm}{logical, deciding whether standardization should be applied to x.
If True, x is standardized using mean and standard deviation}

\item{return.pieces}{if TRUE also returns the pieces of the PAA}
}
\value{
The function returns a SAX representation of x.
}
\description{
The function create SAX symbols for a univariate time series.
The details of this method can be referred to J. Lin, E. Keogh, L. Wei, S. Lonardi.
Experiencing SAX: a novel symbolic representation of time series
Code as in the package TSMining but with a longer alphabet
}
