% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.information.R
\name{get.information}
\alias{get.information}
\title{Information on Processes}
\usage{
get.information(d, dmin, starts, motifs, max_length)
}
\arguments{
\item{d}{original data frame, or data frame with at least a DateTime column}

\item{dmin}{list of not normalized data subsequences}

\item{starts}{starting points of the sequences in the original data}

\item{motifs}{detetcted motifs}

\item{max_length}{length to which the data has been normalized}
}
\value{
returns a data frame with the information necessary for scheduling
}
\description{
Function to extract the "normal" shape from the motifs discovered
}
