% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetProcesses.R
\name{GetProcesses}
\alias{GetProcesses}
\title{Get the process information}
\usage{
GetProcesses(d, x = "median", plot.var = "Median", machine, name,
  event = c("none", "zero", "minimum", "custom"), custom.event = 0,
  norm = TRUE, window.size = 100, alphabet, w = NA, window = NA,
  cr = 1.5, mr = 1.2, mask = 2, iter = 25, epsilon = 0.1, path)
}
\arguments{
\item{d}{original data}

\item{x}{name of the column which includes the data}

\item{plot.var}{define which variable should be plotted}

\item{machine}{name of the machine which is considered}

\item{name}{name of the time series under investigation}

\item{event}{which event search method is used for the subsequences ("none",
zero", "minimum" or "custom")}

\item{custom.event}{custom event start indicator}

\item{norm}{normalise the subsequences to a common length}

\item{window.size}{window size for minimum search}

\item{alphabet}{define the alphabet size for the motif discovery}

\item{w}{word size for motif discovery}

\item{window}{window size for motif discovery}

\item{cr}{count ration for motif discovery (default is 1.5)}

\item{mr}{value used to add other possible members to a motif candidate}

\item{mask}{mask size used for random projection}

\item{iter}{iterations used in random projection algorithm}

\item{epsilon}{epsilom used for threshold in motif discovery}

\item{path}{path to where the plots are printed and the data is saved}
}
\value{
returns complete process motif information and vector with result
}
\description{
Wrapper function for complete motif discovery for energy time series data
}
