% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimum.search.R
\name{minimum.search}
\alias{minimum.search}
\title{Minimum Search}
\usage{
minimum.search(d, x, event = c("none", "zero", "minimum", "custom"),
  custom.event = 0.06, window.size = 100, window)
}
\arguments{
\item{d}{data frame}

\item{x}{name of the column containing the data to be evaluated}

\item{event}{(none, zero, minimum, custom) subsequences are either determined by a
minimum search or thorugh the points where they are zero or another
specified value. If none is selected the subsequences are predefined by
the window length}

\item{custom.event}{the customized value for the event start}

\item{window.size}{indicates the window size for the minimum search}

\item{window}{length of the subsequences if no event search is taking place,
if motif discovery follows it should be same window length}
}
\value{
a list with the starting points and the subsequences
}
\description{
Function to find Minima in the time series that could indicate Process starts
}
