% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.plot.R
\name{motif.plot}
\alias{motif.plot}
\title{Nice plot of the motifs discovered}
\usage{
motif.plot(x, w, window, motif, pdf = TRUE, graphicspath,
  datestart = "01.01.2015 00:00:00", dateformat = "\%d.\%m.\%Y \%H:\%M:\%S",
  dateby = "1 hour", subfigure.1 = FALSE, subfigure.2 = TRUE)
}
\arguments{
\item{x}{the original time series data frame}

\item{w}{word length used for motif discovery}

\item{window}{the window size used for the motif discovery}

\item{motif}{the result of the motif.discover function}

\item{pdf}{is plotted as pdf or in the viewer}

\item{graphicspath}{path where to put the pdf plot if necessary}

\item{datestart}{beginning of date to be show on x axis}

\item{dateformat}{format of the date for the x axis}

\item{dateby}{steps in the date function (default 1 hour)}

\item{subfigure.1}{plot 1 is used is a subfigure, thus the font size is bigger}

\item{subfigure.2}{plot 2 is used is a subfigure, thus the font size is bigger}
}
\value{
nice ggplots
}
\description{
This functions creates a nice ggplot of the disovered motifs highlighted in
the original time series and plotted individually (if wanted)
}
