% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.length.R
\name{normalize.length}
\alias{normalize.length}
\title{Normalize length of time series}
\usage{
normalize.length(dmin, length = c("Max", "Quantile", "Value"), q = 0.9,
  value = 100, plot = TRUE, graphicspath = graphicspath, list = TRUE)
}
\arguments{
\item{dmin}{list with time series of different length}

\item{length}{length of the normalized time series, "Max" for maximum length
"Quantile" for a quantile which should be used, "Value" specify a length}

\item{q}{if length is quantile, then we need q in [0,1]}

\item{value}{of length method is value, need to specify here}

\item{plot}{TRUE/FALSE, should there be a plot of all sequences}

\item{graphicspath}{path to where the normalized time series plot is saved}

\item{list}{(TRUE/FALSE), not sure what I wanted to do there}
}
\value{
list of the length normalized time series
}
\description{
Function to find Minima in the time series that could indicate Process starts
}
