% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.plot.R
\name{pdf.plot}
\alias{pdf.plot}
\title{Function which uses a stored plot and puts it in a pdf}
\usage{
pdf.plot(p, n = "graph", w = 14/2.54, h = 10/2.54)
}
\arguments{
\item{p}{the graph to be saved in a pdf}

\item{n}{name for the pdf file}

\item{w}{width of the pdf file}

\item{h}{height of the pdf file}
}
\value{
pdf file of plot
}
\description{
Function which uses a stored plot and puts it in a pdf
}
