% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.SAX.R
\name{vis.SAX}
\alias{vis.SAX}
\title{SAX plot}
\usage{
vis.SAX(d, x, a, time = "None", path, name = "SAX", normal, word,
  pdf = TRUE)
}
\arguments{
\item{d}{a data frame with one column the time series to be plotted,
or a vector with the time series}

\item{x}{name of the column which includes the data, or new name}

\item{a}{alphabet length for the SAX}

\item{time}{specifiy the name of the column with the time data or set to "None"}

\item{path}{path where the plot should be saved}

\item{name}{name of the plot, default is SAX}

\item{normal}{specifiy the name of the column with the norm data}

\item{word}{complete word}

\item{pdf}{should there be a pdf file of the plot}
}
\value{
returns the SAX plot
}
\description{
Visualize SAX of a time series.
}
